requirements.txt
# Framework Web
flask==3.0.0
flask-cors==4.0.0

# HTTP & API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# IA & ML
google-generativeai==0.3.2
openai==1.6.1

# Serveur Production
gunicorn==21.2.0
gevent==23.9.1

# Données
pandas==2.1.4
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3

# Monitoring (optionnel)
prometheus-client==0.19.0